+++ Android/app/src/main/java/com/cpen321/cloutservices/timeline/model/Restaurant.java
@@ -17,7 +17,7 @@ public class Restaurant {
     private double distanceFromUser;
 
     @SerializedName("lineupTime")
-    private int lineupTime;
+    private double lineupTime;
 
     @SerializedName("coordinates")
     private Coordinates coordinates;
@@ -52,7 +52,7 @@ public class Restaurant {
     }
 
     public int getLineupTime() {
-        return lineupTime;
+        return (int)lineupTime;
     }
 
     public Coordinates getCoordinates() {
